Added user id on db calls and change the list style

+++ grocery-list/app/lib/database.ts
@@ -10,8 +10,19 @@ import {
 import { createKysely } from '@vercel/postgres-kysely'
 import { sql } from 'kysely'
 import { ListVM } from './definitions';
+import { auth } from '@/auth'
 
 const db = createKysely<Database>();
+
+
+export async function getUserId() {
+    const session = await auth();
+
+    if (!session?.user?.id)
+        throw new Error("Usuário não encontrado.")
+
+    return session.user.id
+}
 
 //#region List Functions
 export async function createList(list: NewList) {
@@ -22,41 +33,51 @@ export async function createList(list: NewList) {
 }
 
 export async function updateList(id: string, newValues: ListUpdate) {
+    let userId = await getUserId();
     await db.updateTable('lists')
         .set(newValues)
         .where('id', '=', id)
+        .where('user_id', '=', userId)
         .execute()
 }
 
 export async function deleteList(id: string) {
+    let userId = await getUserId();
     return await db.deleteFrom('lists')
         .where('id', '=', id)
+        .where('user_id', '=', userId)
         .returningAll()
         .executeTakeFirst()
 }
 
 export async function findListById(id: string) {
+    let userId = await getUserId();
     return await db.selectFrom('lists')
         .where('id', '=', id)
+        .where('user_id', '=', userId)
         .selectAll()
         .executeTakeFirst()
 
 }
 
 export async function findAllLists() {
+    let userId = await getUserId();
     let query = db.selectFrom('lists')
-
-    return await query.orderBy('buy_dt', 'desc')
+        .where('user_id', '=', userId)
+        .orderBy('buy_dt', 'desc')
         .selectAll()
-        .execute()
+
+    return await query.execute()
 }
 
 export async function fetchBoughtProductsByList(listId: string) {
+    let userId = await getUserId();
     return await
         db.selectFrom('productlist')
             .innerJoin('products', 'productlist.product_id', 'products.id')
             .innerJoin('lists', 'lists.id', 'productlist.list_id')
             .where('productlist.list_id', '=', listId)
+            .where('productlist.user_id', '=', userId)
             .orderBy(['products.category', 'products.name'])
             .select([
                 'productlist.id as productList_id',
@@ -72,6 +93,8 @@ export async function fetchBoughtProductsByList(listId: string) {
 }
 
 export async function fetchListsWithTotals() {
+    let userId = await getUserId();
+
     return await sql<ListVM>
         `select 
                                 l.id, 
@@ -81,8 +104,9 @@ export async function fetchListsWithTotals() {
                                 sum(pl.quantity * pl.price) as total
                             from lists l 
                             left join productlist pl on pl.list_id  = l.id 
+                            where l.user_id = ${userId}
                             group by l.id `
-        .execute(db)        
+        .execute(db)
 
 }
 //#endregion List Functions
@@ -96,50 +120,71 @@ export async function createProduct(product: NewProduct) {
 }
 
 export async function updateProduct(id: string, newValues: ProductUpdate) {
+    let userId = await getUserId();
+
     await db.updateTable('products')
         .set(newValues)
         .where('id', '=', id)
+        .where('user_id', '=', userId)
         .execute()
 }
 
 export async function deleteProduct(id: string) {
+    let userId = await getUserId();
+
     return await db.deleteFrom('products')
         .where('id', '=', id)
+        .where('user_id', '=', userId)
         .returningAll()
         .executeTakeFirst()
 }
 
 export async function findProductById(id: string) {
+    let userId = await getUserId();
+
     return await db.selectFrom('products')
         .where('id', '=', id)
+        .where('user_id', '=', userId)
         .selectAll()
         .executeTakeFirst()
 }
 
 export async function findProductByName(name: string) {
+    let userId = await getUserId();
+
     return await db.selectFrom('products')
         .where('name', 'ilike', name.trim())
+        .where('user_id', '=', userId)
         .selectAll()
         .executeTakeFirst()
 }
 
 export async function findProductsByCategory(name: string) {
+    let userId = await getUserId();
+
     return await db.selectFrom('products')
         .where('category', 'ilike', name.trim())
+        .where('user_id', '=', userId)
         .selectAll()
         .execute()
 }
 
 export async function findAllProducts() {
+    let userId = await getUserId();
+
     return await db.selectFrom('products')
+        .where('user_id', '=', userId)
         .orderBy('name')
         .selectAll()
         .execute()
 }
 
 export async function findAllCategories() {
+    let userId = await getUserId();
+
     return await db.selectFrom('products')
         .select('category')
+        .where('user_id', '=', userId)
         .orderBy('category')
         .distinct()
         .execute()
@@ -155,9 +200,10 @@ export async function createProductList(productList: NewProductList) {
 }
 
 export async function createBunchProductList(listId: string, products: string[]) {
+    let userId = await getUserId();
     return await db
         .insertInto('productlist')
-        .columns(['list_id', 'product_id', 'quantity', 'price'])
+        .columns(['list_id', 'product_id', 'quantity', 'price', 'user_id'])
         .expression((eb) => eb
             .selectFrom('products')
             .where('id', 'in', products)
@@ -166,30 +212,40 @@ export async function createBunchProductList(listId: string, products: string[])
                 'products.id',
                 eb.lit(0).as('quantity'),
                 eb.lit(0).as('price'),
+                eb.val(userId).as('user_id'),
             ])
         )
         .execute()
 }
 
 export async function updateProductList(id: string, newValues: ProductListUpdate) {
+    let userId = await getUserId();
+
     await db.updateTable('productlist')
         .set(newValues)
         .where('id', '=', id)
+        .where('user_id', '=', userId)
         .execute()
 }
 
 export async function deleteProductList(id: string) {
+    let userId = await getUserId();
+
     return await db.deleteFrom('productlist')
         .where('id', '=', id)
+        .where('user_id', '=', userId)
         .returningAll()
         .executeTakeFirst()
 }
 
 export async function findProductListById(id: string) {
+    let userId = await getUserId();
+
     return await db.selectFrom('productlist')
         .innerJoin('products', 'productlist.product_id', 'products.id')
         .innerJoin('lists', 'productlist.list_id', 'lists.id')
         .where('productlist.id', '=', id)
+        .where('productlist.user_id', '=', userId)
         .select([
             'productlist.id',
             'productlist.quantity',
@@ -205,8 +261,11 @@ export async function findProductListById(id: string) {
 }
 
 export async function fetchProductListById(id: string) {
+    let userId = await getUserId();
+
     return await db.selectFrom('productlist')
-        .where('productlist.id', '=', id)
+        .where('id', '=', id)
+        .where('user_id', '=', userId)
         .select([
             'id',
             'product_id',
@@ -220,6 +279,7 @@ export async function fetchProductListById(id: string) {
 }
 
 export async function findProductNotInList(listId: string) {
+    let userId = await getUserId();
     return await db
         .selectFrom('products')
         .where(({ not, exists, selectFrom }) =>
@@ -235,6 +295,7 @@ export async function findProductNotInList(listId: string) {
             'products.name',
             'products.category',
         ])
+        .where('user_id', '=', userId)
         .orderBy('products.name')
         .execute()
 }
